Check if directory exists when deleting

+++ lib/models/music/playlist.dart
@@ -68,7 +68,9 @@ class Playlist {
   Future<void> delete({bool upload = true}) async {
     var file = File(path);
     var directory = Directory(PathUtils.join(appStorage.playlistsPath, id));
-    await directory.delete(recursive: true);
+    if(await directory.exists()) {
+      await directory.delete(recursive: true);
+    }
     await file.delete();
     if(upload) {
       appWebChannel.deletePlaylist(id: id);
